requirements-tcga.txt
pandas>=2.0
numpy>=1.24
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12
google-cloud-bigquery>=3.11
google-cloud-storage>=2.10
GEOparse>=2.0
